Change regexp to match empty string when filter is enabled
When an option like "show searches" is turned off, this enables a
filter to remove searches from the places model. The existing
regex for this filter matches on the "url" field and ends with ".+",
which means that items with a blank URL will not be matched, even
though they aren't searches. Devices that have not yet been setup can
have a blank URL.

The same issue is fixed for all "url" filters.